Added changes

+++ src/main/java/com/platzi/market/persistence/entity/Cliente.java
@@ -19,7 +19,7 @@ public class Cliente {
 
     private String apellidos;
 
-    private Integer celular;
+    private Long celular;
 
     private String direccion;
 
@@ -29,86 +29,50 @@ public class Cliente {
     @OneToMany(mappedBy = "cliente")
     private List<Compra> compras;
 
-    /**
-     * @return String return the id
-     */
-    public String getId() {
+   public String getId() {
         return id;
     }
 
-    /**
-     * @param id the id to set
-     */
     public void setId(String id) {
         this.id = id;
     }
 
-    /**
-     * @return String return the nombre
-     */
     public String getNombre() {
         return nombre;
     }
 
-    /**
-     * @param nombre the nombre to set
-     */
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
 
-    /**
-     * @return String return the apellidos
-     */
     public String getApellidos() {
         return apellidos;
     }
 
-    /**
-     * @param apellidos the apellidos to set
-     */
     public void setApellidos(String apellidos) {
         this.apellidos = apellidos;
     }
 
-    /**
-     * @return Integer return the celular
-     */
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    /**
-     * @param celular the celular to set
-     */
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
 
-    /**
-     * @return String return the direccion
-     */
     public String getDireccion() {
         return direccion;
     }
 
-    /**
-     * @param direccion the direccion to set
-     */
     public void setDireccion(String direccion) {
         this.direccion = direccion;
     }
 
-    /**
-     * @return String return the correoElectronico
-     */
     public String getCorreoElectronico() {
         return correoElectronico;
     }
 
-    /**
-     * @param correoElectronico the correoElectronico to set
-     */
     public void setCorreoElectronico(String correoElectronico) {
         this.correoElectronico = correoElectronico;
     }
